+++ console/src/components/Services/RemoteSchema/Common/Common.js
@@ -18,6 +18,9 @@ class Common extends React.Component {
     if (valType === 'static') {
       return 'header value';
     }
+    if (valType === 'header') {
+      return 'from header name';
+    }
     return 'env var name';
   }
 
@@ -208,6 +211,7 @@ class Common extends React.Component {
           typeOptions={[
             { display_text: 'Value', value: 'static' },
             { display_text: 'From env var', value: 'env' },
+            { display_text: 'From request header', value: 'header' },
           ]}
           isDisabled={isDisabled}
           placeHolderText={this.getPlaceHolderText.bind(this)}