+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - console: add support for table query and subscription root fields visibility permissions
 - console: fix the BigQuery row-level restrictions on boolean columns
 - console: add support for new DB-to-RS metadata format in old table relationships UI
+- console: add support for header mapping
 
 ## v2.10.0-beta.1